Move files under generate folder
