+++ docs/generate/generate.md
